OK, let's make a real port library.
